+++ src/command/render/crossref.ts
@@ -25,10 +25,6 @@ import { pandocMetadataPath } from "./render-paths.ts";
 import { isMultiFileBookFormat } from "../../project/types/book/book-shared.ts";
 import { projectIsBook } from "../../project/project-context.ts";
 
-export function crossrefFilter() {
-  return resourcePath("filters/crossref/crossref.lua");
-}
-
 export function crossrefFilterActive(options: PandocOptions) {
   return options.format.metadata.crossref !== false;
 }
