+++ src/resources/filters/main.lua
@@ -1,0 +1,321 @@
+-- quarto-pre.lua
+-- Copyright (C) 2020 by RStudio, PBC
+
+-- required version
+PANDOC_VERSION:must_be_at_least '2.13'
+
+-- global state
+preState = {
+  usingBookmark = false,
+  usingTikz = false,
+  results = {
+    resourceFiles = pandoc.List({}),
+    inputTraits = {}
+  },
+  file = nil,
+  appendix = false,
+  fileSectionIds = {},
+  emulatedNodeHandlers = {}
+}
+
+postState = {
+}
+
+-- [import]
+function import(script)
+  local path = PANDOC_SCRIPT_FILE:match("(.*[/\\])")
+  dofile(path .. script)
+end
+
+import("./ast/pandocemulation.lua") -- must come before customnodes.lua
+
+import("./ast/nodeconversion.lua") -- must come before customnodes.lua
+import("./ast/pandocwalk.lua") -- must come before customnodes.lua
+import("./ast/customnodes.lua")
+import("./ast/emulatedfilter.lua")
+import("./ast/parse.lua")
+import("./ast/render.lua")
+import("./ast/runemulation.lua")
+
+import("./common/authors.lua")
+import("./common/base64.lua")
+import("./common/colors.lua")
+import("./common/debug.lua")
+import("./common/error.lua")
+import("./common/figures.lua")
+import("./common/filemetadata.lua")
+import("./common/format.lua")
+import("./common/latex.lua")
+import("./common/layout.lua")
+import("./common/list.lua")
+import("./common/log.lua")
+import("./common/lunacolors.lua")
+import("./common/maporcall.lua")
+import("./common/meta.lua")
+import("./common/options.lua")
+import("./common/pandoc.lua")
+import("./common/paths.lua")
+import("./common/refs.lua")
+import("./common/string.lua")
+import("./common/table.lua")
+import("./common/tables.lua")
+import("./common/theorems.lua")
+import("./common/timing.lua")
+import("./common/url.lua")
+import("./common/validate.lua")
+import("./common/wrapped-filter.lua")
+
+import("./quarto-init/includes.lua")
+import("./quarto-init/resourcerefs.lua")
+
+import("./quarto-post/book.lua")
+import("./quarto-post/delink.lua")
+import("./quarto-post/fig-cleanup.lua")
+import("./quarto-post/foldcode.lua")
+import("./quarto-post/ipynb.lua")
+import("./quarto-post/latexdiv.lua")
+import("./quarto-post/meta.lua")
+import("./quarto-post/ojs.lua")
+import("./quarto-post/responsive.lua")
+import("./quarto-post/reveal.lua")
+import("./quarto-post/tikz.lua")
+
+import("./quarto-finalize/dependencies.lua")
+import("./quarto-finalize/book-cleanup.lua")
+import("./quarto-finalize/mediabag.lua")
+import("./quarto-finalize/meta-cleanup.lua")
+
+import("./authors/authors.lua")
+
+import("./layout/meta.lua")
+import("./layout/width.lua")
+import("./layout/latex.lua")
+import("./layout/html.lua")
+import("./layout/wp.lua")
+import("./layout/docx.lua")
+import("./layout/odt.lua")
+import("./layout/pptx.lua")
+import("./layout/table.lua")
+import("./layout/figures.lua")
+import("./layout/cites.lua")
+import("./layout/columns.lua")
+import("./layout/options.lua")
+import("./layout/columns-preprocess.lua")
+import("./layout/layout.lua")
+
+import("./crossref/index.lua")
+import("./crossref/preprocess.lua")
+import("./crossref/sections.lua")
+import("./crossref/figures.lua")
+import("./crossref/tables.lua")
+import("./crossref/equations.lua")
+import("./crossref/listings.lua")
+import("./crossref/theorems.lua")
+import("./crossref/qmd.lua")
+import("./crossref/refs.lua")
+import("./crossref/meta.lua")
+import("./crossref/format.lua")
+import("./crossref/options.lua")
+import("./crossref/crossref.lua")
+initCrossrefIndex()
+
+import("./quarto-pre/bibliography-formats.lua")
+import("./quarto-pre/book-links.lua")
+import("./quarto-pre/book-numbering.lua")
+import("./quarto-pre/callout.lua")
+import("./quarto-pre/code-filename.lua")
+import("./quarto-pre/content-hidden.lua")
+import("./quarto-pre/engine-escape.lua")
+import("./quarto-pre/figures.lua")
+import("./quarto-pre/hidden.lua")
+import("./quarto-pre/include-paths.lua")
+import("./quarto-pre/input-traits.lua")
+import("./quarto-pre/line-numbers.lua")
+import("./quarto-pre/meta.lua")
+import("./quarto-pre/options.lua")
+import("./quarto-pre/output-location.lua")
+import("./quarto-pre/outputs.lua")
+import("./quarto-pre/panel-input.lua")
+import("./quarto-pre/panel-layout.lua")
+import("./quarto-pre/panel-sidebar.lua")
+import("./quarto-pre/panel-tabset.lua")
+import("./quarto-pre/profile.lua")
+import("./quarto-pre/project-paths.lua")
+import("./quarto-pre/resourcefiles.lua")
+import("./quarto-pre/results.lua")
+import("./quarto-pre/shortcodes-handlers.lua")
+import("./quarto-pre/shortcodes.lua")
+import("./quarto-pre/table-captions.lua")
+import("./quarto-pre/table-colwidth.lua")
+import("./quarto-pre/table-rawhtml.lua")
+import("./quarto-pre/theorems.lua")
+
+-- [/import]
+
+initShortcodeHandlers()
+
+local quartoInit = {
+  { name = "init-readIncludes", filter = readIncludes() },
+  { name = "init-metadataResourceRefs", filter = combineFilters({
+    fileMetadata(),
+    resourceRefs()
+  })},
+}
+
+local quartoAuthors = {
+  { name = "authors", filter = authorsFilter() }
+}
+
+local quartoPre = {
+  -- quarto-pre
+  { name = "pre-parseExtendedNodes", filter = parseExtendedNodes() },
+  { name = "pre-quartoBeforeExtendedUserFilters", filters = make_emulated_user_filters("beforeQuartoFilters") },
+  { name = "pre-bibliographyFormats", filter = bibliographyFormats() }, 
+  { name = "pre-shortCodesBlocks", filter = shortCodesBlocks() } ,
+  { name = "pre-shortCodesInlines", filter = shortCodesInlines() },
+  { name = "pre-tableMergeRawHtml", filter = tableMergeRawHtml() },
+  { name = "pre-tableRenderRawHtml", filter = tableRenderRawHtml() },
+  { name = "pre-tableColwidthCell", filter = tableColwidthCell() },
+  { name = "pre-tableColwidth", filter = tableColwidth() },
+  { name = "pre-hidden", filter = hidden() },
+  { name = "pre-contentHidden", filter = contentHidden() },
+  { name = "pre-tableCaptions", filter = tableCaptions() },
+  { name = "pre-outputs", filter = outputs() },
+  { name = "pre-outputLocation", filter = outputLocation() },
+  { name = "pre-combined-figures-theorems-etc", filter = combineFilters({
+    fileMetadata(),
+    configProfile(),
+    indexBookFileTargets(),
+    bookNumbering(),
+    includePaths(),
+    resourceFiles(),
+    quartoPreFigures(),
+    quartoPreTheorems(),
+    callout(),
+    codeFilename(),
+    lineNumbers(),
+    engineEscape(),
+    panelInput(),
+    panelTabset(),
+    panelLayout(),
+    panelSidebar(),
+    inputTraits()
+  }) },
+  { name = "pre-combined-book-file-targets", filter = combineFilters({
+    fileMetadata(),
+    resolveBookFileTargets(),
+  }) },
+  { name = "pre-quartoPreMetaInject", filter = quartoPreMetaInject() },
+  { name = "pre-writeResults", filter = writeResults() },
+  { name = "pre-projectPaths", filter = projectPaths()},
+}
+
+local quartoPost = {
+  -- quarto-post
+  { name = "post-foldCode", filter = foldCode() },
+  { name = "post-figureCleanupCombined", filter = combineFilters({
+    latexDiv(),
+    responsive(),
+    ipynb(),
+    quartoBook(),
+    reveal(),
+    tikz(),
+    delink(),
+    figCleanup()
+  }) },
+  { name = "post-ojs", filter = ojs() },
+  { name = "post-postMetaInject", filter = quartoPostMetaInject() },
+  { name = "post-renderExtendedNodes", filter = renderExtendedNodes() },
+  { name = "post-userAfterQuartoFilters", filter = make_emulated_user_filters("afterQuartoFilters") },
+}
+
+local quartoFinalize = {
+    -- quarto-finalize
+    { name = "finalize-fileMetadataAndMediabag", filter =
+    combineFilters({
+      fileMetadata(),
+      mediabag()
+    })
+  },
+  { name = "finalize-bookCleanup", filter = bookCleanup() },
+  { name = "finalize-metaCleanup", filter = metaCleanup() },
+  { name = "finalize-dependencies", filter = dependencies() },
+}
+
+local quartoLayout = {
+  { name = "layout-columnsPreprocess", filter = columnsPreprocess() },
+  { name = "layout-columns", filter = columns() },
+  { name = "layout-citesPreprocess", filter = citesPreprocess() },
+  { name = "layout-cites", filter = cites() },
+  { name = "layout-panels", filter = layoutPanels() },
+  { name = "layout-extendedFigures", filter = extendedFigures() },
+  { name = "layout-metaInject", filter = layoutMetaInject() }
+}
+
+local quartoCrossref = {
+  { name = "crossref-initCrossrefOptions", filter = initCrossrefOptions() },
+  { name = "crossref-preprocess", filter = crossrefPreprocess() },
+  { name = "crossref-preprocessTheorems", filter = crossrefPreprocessTheorems() },
+  { name = "crossref-combineFilters", filter = combineFilters({
+    fileMetadata(),
+    qmd(),
+    sections(),
+    crossrefFigures(),
+    crossrefTables(),
+    equations(),
+    listings(),
+    crossrefTheorems(),
+  })},
+  { name = "crossref-resolveRefs", filter = resolveRefs() },
+  { name = "crossref-crossrefMetaInject", filter = crossrefMetaInject() },
+  { name = "crossref-writeIndex", filter = writeIndex() },
+}
+
+local filterList = {}
+
+tappend(filterList, quartoInit)
+tappend(filterList, quartoAuthors) -- FIXME only run when enabled.
+tappend(filterList, quartoPre)
+tappend(filterList, quartoCrossref) -- FIXME only run when enabled.
+tappend(filterList, quartoLayout)
+tappend(filterList, quartoPost)
+tappend(filterList, quartoFinalize)
+
+return run_as_extended_ast({
+  pre = {
+    initOptions()
+  },
+  filters = capture_timings(filterList),
+})
+
+-- The default order of filters will be
+-- quarto-init
+-- quarto-authors
+-- user filters
+-- extension filters
+-- quarto-filters <quarto>
+-- quarto-finalizer
+-- citeproc
+
+-- TODO!!
+-- authors filter toggle using info from authors.ts:authorsFilterActive
+-- crossref filter toggle using info from crossref.ts:crossrefFilterActive
+-- citeproc detection/toggle
+
+--[[ from filters.ts:
+
+// citeproc at the very end so all other filters can interact with citations
+filters = filters.filter((filter) => filter !== kQuartoCiteProcMarker);
+const citeproc = citeMethod(options) === kQuartoCiteProcMarker;
+if (citeproc) {
+  // If we're explicitely adding the citeproc filter, turn off
+  // citeproc: true so it isn't run twice
+  // See https://github.com/quarto-dev/quarto-cli/issues/2393
+  if (options.format.pandoc.citeproc === true) {
+    delete options.format.pandoc.citeproc;
+  }
+
+  quartoFilters.push(kQuartoCiteProcMarker);
+}
+
+]]
